requirements.txt
requests
google-genai
python-dotenv
pillow
pymongo
tweepy
